+++ UmbrellaPipeline/path_finding/node.py
